refactor examples
