+++ examples/src/models/chat_streaming.ts
@@ -1,0 +1,40 @@
+import { CallbackManager } from "langchain/callbacks";
+import { ChatOpenAI } from "langchain/chat_models";
+import { HumanChatMessage } from "langchain/schema";
+
+export const run = async () => {
+  const chat = new ChatOpenAI({
+    maxTokens: 25,
+    streaming: true,
+    callbackManager: CallbackManager.fromHandlers({
+      async handleLLMNewToken(token: string) {
+        console.log({ token });
+      },
+    }),
+  });
+
+  const response = await chat.call([
+    new HumanChatMessage("Tell me a joke."),
+  ]);
+
+  console.log(response);
+  // { token: '' }
+  // { token: '\n\n' }
+  // { token: 'Why' }
+  // { token: ' don' }
+  // { token: "'t" }
+  // { token: ' scientists' }
+  // { token: ' trust' }
+  // { token: ' atoms' }
+  // { token: '?\n\n' }
+  // { token: 'Because' }
+  // { token: ' they' }
+  // { token: ' make' }
+  // { token: ' up' }
+  // { token: ' everything' }
+  // { token: '.' }
+  // { token: '' }
+  // AIChatMessage {
+  //   text: "\n\nWhy don't scientists trust atoms?\n\nBecause they make up everything."
+  // }
+};
